add model entity and controller

+++ server/src/product/product.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { ProductDTO } from './product.dto';
@@ -11,27 +11,39 @@ export class ProductService {
     private productRepository: Repository<ProductEntity>,
   ) {}
 
-  async showAll() {
-    return await this.productRepository.find();
+  async showAll(): Promise<ProductEntity[]> {
+    return this.productRepository.find();
   }
 
-  async create(data: ProductDTO) {
-    const product = await this.productRepository.create(data);
-    await this.productRepository.save(product);
+  async create(data: ProductDTO): Promise<ProductEntity> {
+    const product = this.productRepository.create(data);
+    this.productRepository.save(product);
     return product;
   }
 
-  async showOne(id: string) {
-    return await this.productRepository.findOne({ id });
+  async showOne(id: string): Promise<ProductEntity> {
+    const product = await this.productRepository.findOne({ id });
+    if (!product) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
+    return product;
   }
 
-  async update(id: string, data: Partial<ProductDTO>) {
+  async update(id: string, data: Partial<ProductDTO>): Promise<ProductEntity> {
+    const product = await this.productRepository.findOne({ id });
+    if (!product) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
     await this.productRepository.update({ id }, data);
-    return await this.productRepository.findOne({ id });
+    return this.productRepository.findOne({ id });
   }
 
-  async delete(id: string) {
+  async delete(id: string): Promise<ProductEntity> {
+    const product = await this.productRepository.findOne({ id });
+    if (!product) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
     await this.productRepository.delete({ id });
-    return { deleted: true };
+    return product;
   }
 }
